ai-service/requirements.txt
anthropic==0.39.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
